+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 
 ## 1.0.3
 
-- Initial release of `riverpod_cli`.
-- Added support for `riverpod_cli init` to initialize the project structure with folders and base configuration.
-- Added `riverpod_cli create page:<page_name>` command to generate a feature structure with:
+- Initial release of `river_cli`.
+- Added support for `river_cli init` to initialize the project structure with folders and base configuration.
+- Added `river_cli create page:<page_name>` command to generate a feature structure with:
   - `controllers`
   - `bindings`
   - `views`